chore: Prefetch lib versions
