+++ src/content/content.ts
@@ -11,6 +11,7 @@ import githubIcon from '../assets/icons/github.svg';
 import storybookIcon from '../assets/icons/storybook.svg';
 import {CustomBlockModel} from '../blocks/types';
 import {CustomBlock} from '../blocks/constants';
+import {libs} from '../libs.mjs';
 
 interface CustomPageContent {
     blocks: (Block | CustomBlockModel)[];
@@ -82,47 +83,7 @@ const typedContent: CustomPageContent = {
                 md: 6,
                 lg: 4,
             },
-            items: [
-                {
-                    title: 'UIKit',
-                    text: 'A library with a wide range of easy-to-use components for creating interfaces of any complexity.',
-                    githubId: 'gravity-ui/uikit',
-                    npmId: '@gravity-ui/uikit',
-                    storybookUrl: 'https://preview.yandexcloud.dev/uikit/',
-                },
-                {
-                    title: 'Page&nbsp;constructor',
-                    text: 'A set of stylish, functional blocks to quickly create promo and landing pages.',
-                    githubId: 'gravity-ui/page-constructor',
-                    npmId: '@gravity-ui/page-constructor',
-                    storybookUrl: 'https://preview.yandexcloud.dev/page-constructor/',
-                },
-                {
-                    title: 'ChartKit',
-                    text: 'A library for data visualization, integrated with our design system.',
-                    githubId: 'gravity-ui/chartkit',
-                    npmId: '@gravity-ui/chartkit',
-                },
-                {
-                    title: 'Navigation',
-                    text: 'Compact and extensible navigation panel for complex services.',
-                    githubId: 'gravity-ui/navigation',
-                    npmId: '@gravity-ui/navigation',
-                    storybookUrl: 'https://preview.yandexcloud.dev/navigation/',
-                },
-                {
-                    title: 'i18n',
-                    text: 'Lightweight internationalization helper.',
-                    githubId: 'gravity-ui/i18n',
-                    npmId: '@gravity-ui/i18n',
-                },
-                {
-                    title: 'DashKit',
-                    text: 'Grid component for building interactive dashboards.',
-                    githubId: 'gravity-ui/dashkit',
-                    npmId: '@gravity-ui/dashkit',
-                },
-            ],
+            items: libs,
         },
         {
             type: CustomBlock.CustomBanner,
