+++ src/versions.json
@@ -1,0 +1,1 @@
+[{"title":"@gravity-ui/navigation","version":"0.3.0","date":"24.11.2022"},{"title":"@gravity-ui/i18n","version":"1.0.0","date":"08.09.2022"},{"title":"@gravity-ui/dashkit","version":"3.4.0","date":"16.11.2022"},{"title":"@gravity-ui/chartkit","version":"1.4.3","date":"14.12.2022"},{"title":"@gravity-ui/uikit","version":"3.11.0","date":"14.12.2022"},{"title":"@gravity-ui/page-constructor","version":"1.10.3","date":"15.12.2022"}]
